Move the SSE streaming conversion into serve/stream.rs
